Add minimal OS X support
Previously, the nif_mount function only supported Linux as it used the
the getmntent. This commit adds an alternate implementation of
nif_mount which uses the getmntinfo function available on OS X.

While the getmntinfo returns data about the flags, I haven't
implemented parsing for them here as they don't seem to be used.

Signed-off-by: Steven Danna <steve@chef.io>